requirements.txt
a2a-sdk[http-server]>=0.3.0
google-genai>=1.0.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
httpx>=0.27.0
